Bug 1966196 - Expand EnvironmentCoordinate hops bitwidth to 16 r=nbp
16 bits allows JSOp::EnvCallee to support more than 255 hops, but that means
GetAliasedVar, GetAliasedDebugVar, SetAliasedVar, InitAliasedLexical, and
CheckAliasedLexical also increase their length to make hops 16 bits.

Tests breakpoint-dot-generator.js and bug1368736.js needed to adjust their
bytecode offsets to mach the new operation bytelengths.

Test super-in-nested-eval.js no longer runs into the limit on environment chain
lengths (which was 255 and is now 65535), so I set the maximum depth to 300 to
show the 255 limit can now be broken, and verified that large enough depth will
throw InternalError, rather than crashing, by using a ridiculously large
number.

Differential Revision: https://phabricator.services.mozilla.com/D270389

UltraBlame original commit: 89313c06df5e30c50ab42494e3435b6c69e18840

+++ js/src/frontend/EmitterScope.h
@@ -57,7 +57,7 @@ class MOZ_STACK_CLASS EmitterScope : public Nestable<EmitterScope> {
 #endif
 
   
-  uint8_t environmentChainLength_;
+  uint16_t environmentChainLength_;
